Removed API and moved to Server rendered
