--- SELearning.Api/SELearning.Api.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\SELearning.Core\SELearning.Core.csproj" />
-    <ProjectReference Include="..\SELearning.Infrastructure\SELearning.Infrastructure.csproj" />
-  </ItemGroup>
-
-</Project>
