+++ SELearning.Core/GlobalUsings.cs
@@ -1,0 +1,1 @@
+global using System.ComponentModel.DataAnnotations;
