+++ SELearning.Core/WeatherForecastDTO.cs
@@ -1,0 +1,7 @@
+namespace SELearning.Core;
+
+public record WeatherForecastDTO {
+    public DateTime Date;
+    public int TemperatureC;
+    public string Summary;
+}
