+++ SELearning.Core/WeatherForecastRepository.cs
@@ -1,0 +1,5 @@
+namespace SELearning.Core;
+
+public interface IWeatherForecastRepository {
+    Task<WeatherForecastDTO[]> ReadAsync(DateTime startDate);
+}
