+++ SELearning.Infrastructure/GlobalUsings.cs
@@ -1,0 +1,3 @@
+global using System.ComponentModel.DataAnnotations;
+
+global using SELearning.Core;
