+++ SELearning/GlobalUsings.cs
@@ -1,0 +1,4 @@
+global using SELearning.Services;
+
+global using SELearning.Core;
+global using SELearning.Infrastructure;
