+++ SELearning/SELearning.csproj
@@ -1,10 +1,16 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Identity.Web" Version="1.*" />
     <PackageReference Include="Microsoft.Identity.Web.UI" Version="1.*" />
     <PackageReference Include="Microsoft.Identity.Web.MicrosoftGraph" Version="1.*" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\SELearning.Core\SELearning.Core.csproj" />
+    <ProjectReference Include="..\SELearning.Infrastructure\SELearning.Infrastructure.csproj" />
   </ItemGroup>
 </Project>
