+++ SELearning/_Imports.razor
@@ -7,3 +7,5 @@
 @using Microsoft.JSInterop
 @using SELearning
 @using SELearning.Shared
+@using SELearning.Core
+@using SELearning.Infrastructure
